Fix checker error, try 1.

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -9,7 +9,7 @@
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
 	unsigned int count = 0;
-	unsigned int xor_result;
+	unsigned long int xor_result;
 
 	xor_result = n ^ m;
 	while (xor_result > 0)
